Unify getting agent version for EKS enrollment during manual and auto-discovery flows
